[onert] Move OptimizerCode and OptimizerInfo into ir
This commit moves OptimizerCode and OptimizerInfo into ir.

ONE-DCO-1.0-Signed-off-by: ragmani <ragmani0216@gmail.com>
