+++ runtime/onert/core/include/compiler/train/TrainingInfo.h
@@ -18,8 +18,9 @@
 #define __ONERT_COMPILER_TRAIN_TRAINING_INFO_H__
 
 #include "ir/Index.h"
-#include "exec/train/optimizer/OptimizerCode.h"
 #include "ir/operation/Loss.h"
+#include "ir/train/OptimizerCode.h"
+#include "ir/train/OptimizerInfo.h"
 
 namespace onert
 {
@@ -32,13 +33,6 @@ struct LossInfo
 {
   ir::operation::Loss::Type type;
   // TODO Add members for loss
-};
-
-struct OptimizerInfo
-{
-  exec::train::optimizer::OptimizerCode optim_code;
-  float learning_rate;
-  // TODO Add properties
 };
 
 class TrainingInfo
@@ -55,12 +49,15 @@ public:
   void setBatchSize(const uint32_t batch_size) { _batch_size = batch_size; }
   const LossInfo &lossInfo() const { return _loss_info; }
   void setLossInfo(const LossInfo &loss_info) { _loss_info = loss_info; }
-  const OptimizerInfo &optimizerInfo() const { return _optimizer_info; }
-  void setOptimizerInfo(const OptimizerInfo &optimizer_info) { _optimizer_info = optimizer_info; }
+  const ir::train::OptimizerInfo &optimizerInfo() const { return _optimizer_info; }
+  void setOptimizerInfo(const ir::train::OptimizerInfo &optimizer_info)
+  {
+    _optimizer_info = optimizer_info;
+  }
 
 private:
   LossInfo _loss_info;
-  OptimizerInfo _optimizer_info;
+  ir::train::OptimizerInfo _optimizer_info;
   uint32_t _batch_size;
 };
