+++ runtime/onert/core/include/ir/train/OptimizerCode.h
@@ -14,29 +14,37 @@
  * limitations under the License.
  */
 
-#include "exec/train/optimizer/OptimizerCode.h"
+#ifndef __ONERT_IR_TRAIN_OPTIMIZER_CODE_H__
+#define __ONERT_IR_TRAIN_OPTIMIZER_CODE_H__
 
-#include <unordered_map>
+#include <functional>
+#include <stdint.h>
+#include <string>
 
 namespace onert
 {
-namespace exec
+namespace ir
 {
 namespace train
 {
-namespace optimizer
-{
 
-std::string toString(OptimizerCode code)
+enum class OptimizerCode
 {
-  static const std::unordered_map<OptimizerCode, const char *> map{
-    {OptimizerCode::Invalid, "Invalid"},
-    {OptimizerCode::SGD, "SGD"},
-    {OptimizerCode::Adam, "Adam"}};
-  return map.at(code);
-}
+  Invalid, //< Invalid
+  SGD,     //< SGD optimizer
+  Adam     //< Adam optimizer
+};
 
-} // namespace optimizer
+/**
+ * @brief Convert the optimizer code to the name
+ *
+ * @param opcode The optimizer code
+ * @return The name of the optimizer
+ */
+std::string toString(OptimizerCode opcode);
+
 } // namespace train
-} // namespace exec
+} // namespace ir
 } // namespace onert
+
+#endif // __ONERT_IR_TRAIN_OPTIMIZER_CODE_H__
