+++ runtime/onert/core/src/compiler/train/TrainingCompiler.cc
@@ -255,11 +255,11 @@ std::shared_ptr<CompilerArtifact> TrainingCompiler::compile(void)
   // TODO Set properties of optimizer
   std::shared_ptr<exec::train::optimizer::Optimizer> optimizer;
   const auto &optim_info = _training_info.optimizerInfo();
-  if (optim_info.optim_code == exec::train::optimizer::OptimizerCode::SGD)
+  if (optim_info.optim_code == ir::train::OptimizerCode::SGD)
     optimizer = std::make_shared<exec::train::optimizer::SGD>(optim_info.learning_rate);
   else
     throw std::runtime_error("Invalid optimizer type, " +
-                             exec::train::optimizer::toString(optim_info.optim_code));
+                             ir::train::toString(optim_info.optim_code));
 
   /*************************************************************
    *  Backend independent analysis & optimization phase finished
