+++ runtime/onert/core/src/ir/train/OptimizerCode.cc
@@ -14,40 +14,26 @@
  * limitations under the License.
  */
 
-#ifndef __ONERT_EXEC_TRAIN_OPTIMIZER_OPTIMIZER_CODE_H__
-#define __ONERT_EXEC_TRAIN_OPTIMIZER_OPTIMIZER_CODE_H__
+#include "ir/train/OptimizerCode.h"
 
-#include <functional>
-#include <stdint.h>
-#include <string>
+#include <unordered_map>
 
 namespace onert
 {
-namespace exec
+namespace ir
 {
 namespace train
 {
-namespace optimizer
+
+std::string toString(OptimizerCode code)
 {
+  static const std::unordered_map<OptimizerCode, const char *> map{
+    {OptimizerCode::Invalid, "Invalid"},
+    {OptimizerCode::SGD, "SGD"},
+    {OptimizerCode::Adam, "Adam"}};
+  return map.at(code);
+}
 
-enum class OptimizerCode
-{
-  Invalid, //< Invalid
-  SGD,     //< SGD optimizer
-  Adam     //< Adam optimizer
-};
-
-/**
- * @brief Convert the optimizer code to the name
- *
- * @param opcode The optimizer code
- * @return The name of the optimizer
- */
-std::string toString(OptimizerCode opcode);
-
-} // namespace optimizer
 } // namespace train
-} // namespace exec
+} // namespace ir
 } // namespace onert
-
-#endif // __ONERT_EXEC_TRAIN_OPTIMIZER_OPTIMIZER_CODE_H__
